convert game into svelte component (wip)
