+++ .eslintignore
@@ -4,10 +4,12 @@ node_modules
 /.svelte-kit
 /package
 .env
-.env.*
+.env.\*
 !.env.example
+src/lib/components/error-game/js
 
 # Ignore files for PNPM, NPM and YARN
+
 pnpm-lock.yaml
 package-lock.json
 yarn.lock
