+++ .prettierignore
@@ -6,6 +6,7 @@ node_modules
 .env
 .env.*
 !.env.example
+src/lib/components/error-game/js
 
 # Ignore files for PNPM, NPM and YARN
 pnpm-lock.yaml
