--- src/lib/components/ErrorGame/capy-styles.css
@@ -1,38 +1,0 @@
-* {
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-}
-
-#game-canvas {
-    border: 5px solid black;
-    position: absolute;
-    /* Center the canvas in the center of the page */
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-
-    /* Setting max width and height so that we can always
-    see the entire canvas regardless of width or height */
-
-    max-width: 100%;
-    max-height: 100%;
-}
-
-#mobileButton {
-    display: none;
-}
-
-@media (max-width: 748px) {
-    #mobileButton {
-        display: block;
-        width: 100px;
-        height: 80px;
-        position: absolute;
-        padding: 10px;
-        transform: translateX(-50%);
-        bottom: 100px;
-        left: 50%;
-        background-color: bisque;
-    }
-}
