--- src/lib/components/ErrorGame/capy.html
@@ -1,18 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>CapyRun</title>
-    <link rel="stylesheet" href = 'capy-styles.css'>
-    <link rel="preconnect" href="https://fonts.googleapis.com">
-    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-    <link href="https://fonts.googleapis.com/css2?family=Dela+Gothic+One&family=Luckiest+Guy&family=Pixelify+Sans:wght@500&display=swap" rel="stylesheet">
-</head>
-<body>
-    <canvas id="game-canvas"></canvas>
-    <button id="mobileButton">Jump</button>
-
-    <script type="module" src="main.js"></script>
-</body>
-</html>
