+++ static/assets/error-game/beeImage.png
