+++ static/assets/error-game/fireEffect.gif
