+++ static/assets/error-game/hedgehogImage.png
