+++ static/assets/error-game/wizardImage.png
