+++ static/assets/error-game/wolfImage.png
